particle filter stage finished and working (with graphical output)

+++ DVSTracker/particle.cpp
@@ -1,11 +1,11 @@
 #include "particle.h"
 
-Particle::Particle(float posX, float posY, float uncert, float w, double ts)
+Particle::Particle(float posX, float posY, float uncert, int w, double ts)
 {
     set(posX,posY,uncert,w,ts);
 }
 
-void Particle::set(float posX, float posY, float uncert, float w, double ts){
+void Particle::set(float posX, float posY, float uncert, int w, double ts){
     x = posX;
     y = posY;
     uncertainty = uncert;
