+++ DVSTracker/particle.h
@@ -4,12 +4,12 @@
 class Particle
 {
 public:
-    Particle(float posX = 0, float posY = 0, float uncert = 0, float w = 0, double ts = 0);
-    void set(float posX, float posY, float uncert, float w, double ts);
+    Particle(float posX = 0, float posY = 0, float uncert = 0, int w = 0, double ts = 0);
+    void set(float posX, float posY, float uncert, int w, double ts);
 
     float x,y;
     float uncertainty;
-    float weight;
+    int weight;
     double timeStamp;
 };
